+++ tests/Table/GridTest.php
@@ -9,26 +9,10 @@ use Orchestra\Html\Table\Grid;
 class GridTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * Application instance.
-     *
-     * @var \Illuminate\Container\Container
-     */
-    protected $app = null;
-
-    /**
-     * Setup the test environment.
-     */
-    public function setUp()
-    {
-        $this->app = new Container;
-    }
-
-    /**
      * Teardown the test environment.
      */
     public function tearDown()
     {
-        unset($this->app);
         m::close();
     }
 
@@ -39,8 +23,8 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testInstanceOfGrid()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
+        $app = new Container;
+        $app['config'] = $config = m::mock('\Illuminate\Config\Repository');
 
         $config->shouldReceive('get')->once()
             ->with('orchestra/html::table', array())->andReturn(array(
@@ -71,14 +55,9 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testWithMethod()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
+        $stub = new Grid($this->getContainer());
 
         $mock = array(new Fluent);
-        $stub = new Grid($app);
         $stub->with($mock, false);
 
         $refl     = new \ReflectionObject($stub);
@@ -90,7 +69,7 @@ class GridTest extends \PHPUnit_Framework_TestCase
         $model->setAccessible(true);
         $paginate->setAccessible(true);
 
-        $this->assertEquals($mock, $rows->getValue($stub)->data);
+        $this->assertEquals($mock, $stub->rows());
         $this->assertEquals($mock, $model->getValue($stub));
         $this->assertFalse($paginate->getValue($stub));
         $this->assertTrue(isset($stub->model));
@@ -104,17 +83,15 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testWithMethodGivenPaginatorInstance()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
+        $expected = array('foo');
+        $stub = new Grid($this->getContainer());
 
         $model = m::mock('\Illuminate\Pagination\Paginator');
-        $model->shouldReceive('getItems')->once()->andReturn(array('foo'));
+        $model->shouldReceive('getItems')->once()->andReturn($expected);
 
-        $stub = new Grid($app);
         $stub->with($model);
+
+        $this->assertEquals($expected, $stub->rows());
     }
 
     /**
@@ -125,17 +102,15 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testWithMethodGivenModelBuilderInstance()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
+        $expected = array('foo');
+        $stub = new Grid($this->getContainer());
 
         $model = m::mock('\Illuminate\Database\Eloquent\Builder')->makePartial();
-        $model->shouldReceive('paginate')->once()->andReturn(array('foo'));
+        $model->shouldReceive('paginate')->once()->andReturn($expected);
 
-        $stub = new Grid($app);
         $stub->with($model);
+
+        $this->assertEquals($expected, $stub->rows());
     }
 
     /**
@@ -146,17 +121,15 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testWithMethodGivenArrayableInterfaceInstance()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
+        $expected = array('foo');
+        $stub = new Grid($this->getContainer());
 
         $model = m::mock('\Illuminate\Support\Contracts\ArrayableInterface');
-        $model->shouldReceive('toArray')->once()->andReturn(array('foo'));
+        $model->shouldReceive('toArray')->once()->andReturn($expected);
 
-        $stub = new Grid($app);
         $stub->with($model);
+
+        $this->assertEquals($expected, $stub->rows());
     }
 
     /**
@@ -167,16 +140,13 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testWithMethodThrowsAnException()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
+        $stub = new Grid($this->getContainer());
 
         $model = 'Foo';
 
-        $stub = new Grid($app);
         $stub->with($model, false);
+
+        $stub->rows();
     }
 
     /**
@@ -186,13 +156,7 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testLayoutMethod()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
+        $stub = new Grid($this->getContainer());
 
         $refl = new \ReflectionObject($stub);
         $view = $refl->getProperty('view');
@@ -215,14 +179,9 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testOfMethod()
     {
-        $me  = $this;
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
+        $me   = $this;
+        $stub = new Grid($this->getContainer());
 
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
         $stub->with(array(
             new Fluent(array('foo1' => 'Foo1')),
         ), false);
@@ -290,14 +249,94 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testOfMethodThrowsException()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
+        $stub = new Grid($this->getContainer());
         $stub->of('id');
+    }
+
+    /**
+     * Test Orchestra\Html\Table\Grid::searchable() method.
+     *
+     * @test
+     */
+    public function testSearchableMethod()
+    {
+        $attributes = array('email', 'fullname');
+        $app = $this->getContainer();
+
+        $app['request'] = $request = m::mock('\Illuminate\Http\Request');
+
+        $request->shouldReceive('input')->once()->with('q')->andReturn('orchestra*');
+
+        $stub = m::mock('\Orchestra\Html\Table\Grid[setupWildcardQueryFilter]', array($app))
+                    ->shouldAllowMockingProtectedMethods();
+
+        $model = m::mock('\Illuminate\Database\Query\Builder');
+
+        $stub->shouldReceive('setupWildcardQueryFilter')->once()->with($model, 'orchestra*', $attributes)->andReturnNull();
+
+        $stub->with($model);
+
+        $this->assertNull($stub->searchable($attributes));
+    }
+
+    /**
+     * Test Orchestra\Html\Table\Grid::searchable() method
+     * throws exception when model is not a query builder.
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSearchableMethodThrowsException()
+    {
+        $attributes = array('email', 'fullname');
+
+        $stub = new Grid($this->getContainer());
+
+        $stub->with('Foo');
+
+        $stub->searchable($attributes);
+    }
+
+    /**
+     * Test Orchestra\Html\Table\Grid::sortable() method.
+     *
+     * @test
+     */
+    public function testSortableMethod()
+    {
+        $app = $this->getContainer();
+        $app['request'] = $request = m::mock('\Illuminate\Http\Request');
+
+        $request->shouldReceive('input')->once()->with('sort')->andReturn('email')
+            ->shouldReceive('input')->once()->with('order')->andReturn('desc');
+
+        $stub = m::mock('\Orchestra\Html\Table\Grid[setupBasicQueryFilter]', array($app))
+            ->shouldAllowMockingProtectedMethods();
+
+        $model = m::mock('\Illuminate\Database\Query\Builder');
+
+        $stub->shouldReceive('setupBasicQueryFilter')->once()
+            ->with($model, array('sort' => 'email', 'order' => 'desc'))->andReturnNull();
+
+        $stub->with($model);
+
+        $this->assertNull($stub->sortable());
+    }
+
+    /**
+     * Test Orchestra\Html\Table\Grid::sortable() method
+     * throws exception when model is not a query builder.
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSortableMethodThrowsException()
+    {
+        $attributes = array('email', 'fullname');
+
+        $stub = new Grid($this->getContainer());
+
+        $stub->with('Foo');
+
+        $stub->sortable($attributes);
     }
 
     /**
@@ -307,13 +346,7 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testAttributesMethod()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
-
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
+        $stub = new Grid($this->getContainer());
 
         $refl       = new \ReflectionObject($stub);
         $attributes = $refl->getProperty('attributes');
@@ -338,13 +371,8 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testMagicMethodCallThrowsException()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
+        $stub = new Grid($this->getContainer());
 
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
         $stub->invalidMethod();
     }
 
@@ -356,13 +384,8 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testMagicMethodGetThrowsException()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
+        $stub = new Grid($this->getContainer());
 
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
         $stub->invalidProperty;
     }
 
@@ -374,13 +397,8 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testMagicMethodSetThrowsException()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
+        $stub = new Grid($this->getContainer());
 
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
         $stub->invalidProperty = array('foo');
     }
 
@@ -392,13 +410,8 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testMagicMethodSetThrowsExceptionValuesNotAnArray()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
+        $stub = new Grid($this->getContainer());
 
-        $config->shouldReceive('get')->once()
-            ->with('orchestra/html::table', array())->andReturn(array());
-
-        $stub = new Grid($app);
         $stub->attributes = 'foo';
     }
 
@@ -410,13 +423,24 @@ class GridTest extends \PHPUnit_Framework_TestCase
      */
     public function testMagicMethodIssetThrowsException()
     {
-        $app = $this->app;
-        $app['config'] = $config = m::mock('Config');
+        $stub = new Grid($this->getContainer());
+
+        isset($stub->invalidProperty) ? true : false;
+    }
+
+    /**
+     * Get app container.
+     *
+     * @return Container
+     */
+    protected function getContainer()
+    {
+        $app = new Container;
+        $app['config'] = $config = m::mock('\Illuminate\Config\Repository');
 
         $config->shouldReceive('get')->once()
             ->with('orchestra/html::table', array())->andReturn(array());
 
-        $stub = new Grid($app);
-        isset($stub->invalidProperty) ? true : false;
+        return $app;
     }
 }